backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
alembic>=1.13.0
psycopg[binary]>=3.1.0
pydantic-settings>=2.0.0
httpx>=0.26.0
python-multipart>=0.0.6
openai>=1.0.0
